igs_backend/requirements.txt
asgiref==3.8.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
Django==4.2.5
django-cors-headers==4.6.0
django-crontab==0.7.1
django-environ==0.11.2
django-filter==24.3
django-sslserver==0.22
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
gunicorn==20.0.2
idna==3.10
inflection==0.5.1
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.4
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tenacity==9.1.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.5.0
